typing: apply typehints from `IMatcher` protocol class to concrete matchers
This was done with `merge-pyi` after 1) tweaking `interfaces._basetypes.py` to
redefine `HgPathT`, `MatcherT`, and `UserMsgT` to use `typing.NewType()` instead
of a simple alias, and 2) rename each concrete matcher class to
`IMatcher(Protocol)`.  `merge-pyi` then copied over the hints to each class, but
botched the formatting (which `black` cleaned up`), and prefixed `Union` and
`Optional` with `typing.` (which were manually converted to `|` and `| None`
respectively.

For this procedure, pytype 2023.11.21 was used.  (I'm a little surprised that it
processed every class, and not just the last one, but that was convenient.)

--HG--
extra : rebase_source : 078485981bb0f7bc7f893690d5ac507288828fb4
extra : topic : matcher-typehints