typing: preserve the signature of `util.lrucachefunc` wrapped methods
I stumbled across `functools.lru_cache` when trying to figure out how to
typehint `util.propertycache` (since anything with that decoration is seen as
`Any` by pytype), and thought it sounded familiar.  Maybe we can replace the
custom implementation with this.  However, it doesn't appear to have typehints
in the Python 3.11 code that PyCharm shows locally, so continuing with this for
now is reasonable.

Now `_getflog` and `_revsingle` in the fastannotate extension have the same
signature as the method they wrap (instead of just being `Callable`).  The other
places this wraps a method get transformed to `Callable[[Any], Any]` (which is
basically the same thing), but those will get updated automatically when the
wrapped methods are given typehints.

I tried defining a `GenericFuncT` in `interfaces.types` to avoid all of this
boilerplate defining and importing, but ended up getting errors:

    ParseError: First argument to Callable must be a list of argument types (got NamedType(name='_P'))

So I guess this only works when defined on a function itself, and we're stuck
with the boilerplate.

This sort of decorator typehinting will help a lot with the extension
decorators, and likely some other internal stuff.  The descriptor typehinting
(like `util.propertycache`) doesn't seem to work, and even got stuck in an
infinite CPU burning loop with pytype 2024.10.11.

--HG--
extra : rebase_source : 6bf78de6bcb05775bb7683091c167a2731c57503
extra : topic : lru-typing