typing: add typehints around `scmutil.revsingle()`
Also add hints through the things it calls until we arrive at documentation for
what these things are.  This is very narrowly targeted because this is meant to
prove typehints for a decorator on this function are working as expected later
in this series.

--HG--
extra : rebase_source : 5fc1949520bec104ec5f518e2fbadba232606fa4
extra : topic : lru-typing