interfaces: correct the signature of `repository.anyrevs()`
Looking at the code for the implementation class, it's obvious that it should be
a list of bytes, not a string.  I guess it went undetected because the arg
wasn't typed on the interface at all.

--HG--
extra : rebase_source : 1dbfcfc91f95d211aaf81e40970f11bfd85eaa24
extra : topic : lru-typing