subrepo: make the `create` arg of `hgsubrepo._init()` a kwarg
The next commit will preserve the signatures of methods that are decorated with
`@annotatesubrepoerror`.  But doing that caused pytype to complain about a spot
that was invoking this with a kwarg style, so we fix it first:

    File "/home/mharbison/projects/mercurial/mercurial-devel/mercurial/subrepo.py", line 802, in _get:
        Invalid keyword argument create to function hgsubrepo._initrepo [wrong-keyword-args]
             Expected: (self, _, _, _)
      Actually passed: (self, _, _, create)

Creating the subrepo on disk is the exception case, and all callers are supplying
a value now, so leave the kwarg style in place for readability.

--HG--
extra : _rewrite_noise : 66456eab85586d733e475da5bd6842efef528cd9
extra : topic : typing-callable2