revlog: enforce that filelog get the `writable` parameter passed
Now that all caller are updated, we can enforce this and move to the next step
of the plan. For now, the enforcement is just a warning to leave some room for
extension to adapt. We will move to hard enforcement in the next version.

--HG--
extra : rebase_source : df903249c4de894c08490ee81089146547897102
extra : topic : filelog-rw