filelog: move the radix computation at the store level
Previously the filelog was responsible for constructing its radix. This change
makes the store do it instead, and pass the radix to the filelog.

The radix is what we use to derive paths for storing revlogs. For example:

- file in repo: foo/Bar.txt
- radix:        data/foo/Bar.txt
- index file:   data/foo/Bar.txt.i
- encoded:      data/foo/_bar.txt.i
- final:        .hg/store/data/foo/_bar.txt.i

This change is in preparation for the file index. Currently the radix is always
formed by prepending "data/", but in the future we will store radixes in the
file index. This will remove the need for encoding (or rather, we can bake the
encoding into the stored radix). And it will allow us to change where revlog
files are stored without requiring a format change every time.

--HG--
extra : amend_source : 248b2d98999ab6f70f1786ee763ff770ed5df648
extra : rebase_source : 16be95aae94305bb1d0a84b72a6b38520f9a6420
extra : topic : file-index.ready