rust: add an "update from clean" fast-path
We now detect if none of the tracked files are dirty (thus "from clean"), and
do the update in Rust, which should be faster in cases with a lot of updates.

This is still experimental since it is likely to require more performance work.

--HG--
extra : rebase_source : 02c94b63772f5a8f5ef3867ebbf5a3099bceb107
extra : topic : rust.update-from-clean