delta-search: search optimized delta within parent chain
When "delta info" flags are used, we can safely pick arbitrary base for
non-snapshot delta (because they won't be confused as snapshot). So we can use
delta folding to cheaply check if it would be better to bypass the parent and do
a delta lower into that parent chain.

This allow to bypass "redundant" delta that modify similar area of the code. For
content when this happens a lot (like manifest, or some file with specific
patterns), this help to shortening average delta chain by producing a more branching
delta tree.

This the resulting delta tree "organically" reduce the presence of "redundant"
delta information, it speed up content recovery and reduce the amount of
snapshot we have to do, speeding up encoding operation too and reducing storage
footprint.

The criteria for "a better snapshot" is current simple, the new delta should not
be more that X time bigger than the initial one (defaulting to ×1.05).

I am sure we could could further improvement le logic by making criteria for
"better delta" more advanced, but let's start "simple".

We currently only do it for parent delta we actually test, and some useful base
might have been too aggressively pre-filtered. However we cannot simple drop
this pre-filtering as it is very useful in various case (e.g. if one of the
parent is much larger than the content we try to encode).

The associate config is not perfect, but all this is experimental anyway, so we
can smooth this out later. Especially once we experimented more with it.

--HG--
extra : rebase_source : 731385156dc66608cf1b1cb4293ab04c06b7019c
extra : topic : delta-folding