delta-search: skip validation during repoupgrade too
Same rational as the previous changeset and as b58bc4fec521, we don't validate
content when we reuse the delta, so there is no point in validating it when we
have a full snapshot. Since sha-1 checksum can amoutn for ⅔ of the time we spend
cloning some revlog, this provides a nice speedup.

--HG--
extra : rebase_source : d42b4d92dd97d6603213031d7a6061467b1bf82b
extra : topic : skip-validate