typing: keep using Optional for type outside type annotation
the `Foo | None` annotation is not supported on Python 3.9. This is not a
problem for type annotation as `from __future__ import annotations` mean they
are not actually process during normal runtime.

However for Type declaration passed as argument, we need to keep using Optional
as long as we support Python 3.9.

--HG--
extra : topic : fix-typing