tests: stabilize test-branch-change.t on Windows
MSYS was getting involved here and mangling the `::`, which ended up excluding
the working directory parent.  How do I know?  Because when I tried to add `+ .`
to the expression, the command flipped out with a parse error:

    $ hg glog -r '. + (nodefromfile("../not-stable.nodes")::)'
      hg: parse error at 40: syntax error in revset '. + (nodefromfile("..\not-stable.nodes");)'
      (. + (nodefromfile("..\not-stable.nodes");)

Not sure why it did that, since `::` is used in other places.  It must have been
because it was detecting a file path, because getting rid of the directory
separator fixes it.  (Why it would decided that it needs a Windows path
separator when it sees a unix style path is a mystery.)

--HG--
branch : stable
extra : rebase_source : 6f7f88c5a738ca786046cea9d4cde504efb5599a
extra : topic : fix-test-branch-change-test