rust-update: replace the use of generic `create` with `create_working_copy`
We need to do a lot less when looking at the working copy, so let's not
use the same logic.

In theory we should be using the VFS trait and using this for the
implementation of a `WorkingCopyVFS`, but this is due for next cycle.

--HG--
branch : stable
extra : rebase_source : 6498976fa468993f715c813f0ed451563fe9b1a2
extra : topic : rust.update.faster