fileindex: use mtime caching for docket
This makes the fileindexstore cache the fileindex object based on the mtime of
the docket file. This way, if a transaction doesn't write to the file index,
it will not have to be reloaded after repo.invalidate(). "Reloading" is just a
few mmaps so it's not that expensive, but it seems worthwhile to follow the
same pattern as other @storecache files.

--HG--
extra : rebase_source : 09ce401889ccfa32743181f24114167f182ec507
extra : topic : file-index