pyo3: use the lazy ancestors iterator from hg-core, not from vcsgraph
The one in hg-core is much faster because it uses a bitmap instead of
a hash table.

I think updating vcsgraph to use the bitmap would also be a good idea,
but I don't know how to do that and if we're maintaining a fork of it in
mercurial tree anyway, why not use it?

--HG--
extra : rebase_source : 4ece35e481218f570089247a0d34b7598af931c5
extra : topic : pyo3-ancestors-use-hg-not-vcg