contrib: remove outdated Vagrantfile and related scripts
The point of this collection of files was to easily spin up a VM that could be
used to run Mercurial test suite. It was nice to have, since Vagrant can use
actual virtual machine engines like VirtualBox as opposed to e.g. Docker, which
wasn't/isn't always available or practical to use on certain OSes.

It was contributed to the project in 2014 and updated a couple of times, with
the last update done in 2019 (bumping Debian version to Buster).

I have tried to update the setup to use the current Debian version, but didn't
really have success building Mercurial inside the VM for some reason (it wasn't
clear to me why, the process just crashed when building C extensions).

In any case, considering the amount of provisioning required it's not too
difficult to spin up a virtual machine (either using an automated solution like
Vagrant, or by hand) to run Mercurial tests, and set it up according to
something more up-to-date like contrib/heptapod-ci.yml.

The reasons to remove Vagrant support now:

1. it hasn't been updated in a while;
2. we currently have CI that can be used by any contributor;
3. supposedly Docker support is better nowadays than it was back in the day;
4. the company behind Vagrant is not very committed to the open source version
   of its software apparently [1];

[1] https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=1049999

--HG--
extra : topic : rm-vagrant