dispatch: always chdir back in _getlocal
We wrap the section that `chdir` in a try except to ensure we always change back
to the old `cwd` before exiting the function.

We should probably entirely skip the chdir logic for this function doing config
loading only, but this is out of scope of this function.

--HG--
branch : stable
extra : rebase_source : 57f7d3507e8d913144109ca5f509bc3e08005a84
extra : topic : cmdserver-chdir