typing: add some trivial typehints around the config object `remap` arg
I'm very specifically not adding typehints to most of this class because I've
got a good chunk of that done locally, and there are some very thorny issues
around where a value is actually optional that need to be resolved.  This was
easy enough to adapt from the local changes, and is related to the next round of
typehints around the `section` args, which was a distraction from the actual bug
I'm trying to fix.  So we'll revisit this later.

--HG--
extra : rebase_source : 86b7b48ba5c3e5e032bbfb290bfe387a221e8821
extra : topic : webconf-path-validation