typing: add some typehints around the config object `sections` arg
I was going to add validation to the `hg serve --web-conf` file content, and saw
that code passes a list of bytes, which was not exactly what I expected, so I
thought I'd add the typehints and see if there's any fallout.  It's still not
clear what the best type is, so this is an alias to make it easier to tweak if
needed.

--HG--
extra : rebase_source : ae7c05828f72c93ba72e3be93b16eb058f7a6fec
extra : topic : webconf-path-validation