tests: demonstrate the current behavior when `--web-conf` paths are invalid
The current behavior is inconsistent- if any of filesystem paths in the config
are invalid, and also have a glob or recursive glob, the serve aborts before the
process has a chance to background.  If the mapped filesystem path is invalid,
but doesn't use a glob, that isn't noticed until the client gets a 500 upon
accessing it, and the server side error log contains a ParseError.  We should
probably abort in all cases, to make the problem more clear.

Note that in the buggy case, the config is slightly different in that the repo
is mapped away from the server root.  That allows `/` to be shown to work, and
the error is on the repo path.  The prior cases don't matter because the server
dies upon launch, and can't be queried.

Each of the buggy sub-cases is a separate spawn of the server, client access,
and then killing the server, in order to minimize the test changes that goes
with the fix.

--HG--
extra : rebase_source : 09ad4500100ee7a5dd933d801e8220c98e1f7271
extra : topic : webconf-path-validation