revlog: cache uncompressed chunk for revision as we add them
If we have the uncompressed chunk from delta search, we can cache them as we go
instead of having to read and uncompress them later. This is especially true
when processing a bundle from a non-delta-info source using delta-folding. The
delta folding process will always try a folding the delta to the parent, with
the delta of that parent. So we better have it in cache already.

In pratice, this yield a small win (about -3%) in the heavier workload (like
pulling things within mozilla or netbeans), and the small slow down we see in
some case (about 1%) are for much faster case were the difference in absolute
time is really small. In addition, this such effect are more visible in the
non-rust usecase where accessing chunk from disk is "cheaper" while the python
implementation of the uncompressed LRU cache is slower.

--HG--
extra : rebase_source : c95509e0e5ae59a38fe0ce5f18546a008beca207
extra : topic : pre-warm-chunk-cache