dispatch: use _get_cwd in _dispatch
In bb47f0e57ef3, we extracted the `_get_cwd` logic in a function to be able to reuse it in
`_dispatch`, then in 92c104b594c3, we forgot to use it.

So we now uses it.

--HG--
extra : rebase_source : a85a8d8966d63822127a74e89ab4c98beb2d0d85
extra : topic : cycle-breaking