dispatch: pass the dispatcher function down to the command server
Instead of importing dispatch, we now have a way to signal a command need to be
passed the dispatch function and we move it down the chain down to the command
server.

This kill two extra import cycle flag by the import-checker and solve the 5
modules cycle for pytype.

So good progress again.

--HG--
extra : rebase_source : 38802d8994610bc34519b01d9500940bb6612077
extra : topic : cycle-breaking