pyproject: use old license format and avoid rushed setuptools upgrade
5392f4fd6764 introduced SPDX spelling of the pyproject.toml license expression.
The SPDX spelling is a good and simple change that seems to be low risk and
doesn't need further justification.

But that changeset also moved to the new PEP 639 license format where 'license'
is a string and 'license-files' is a list. Support for that format was "just"
introduced in setuptools 77, released Mar 19 2025. That part of the change thus
broke installing Mercurial with older setuptools versions, apparently without
being aware of it, and without considering the impact. That seems like a rushed
upgrade.

16d650ea972f noticed the problem without getting to the root cause, and
worked around it by enforcing an even more recent setuptools version.

This means that for example the current Fedora 42 can not be upgraded to
latest Mercurial. (That is annoying, but not a big problem: The major upgrade
to Mercurial 7.1 will only happen in Fedora 43 where setuptools will get a major
upgrade too.)

Still, the better and more careful approach would be to wait a bit before
deliberately enforcing a new setuptools version.

I suggest rolling the risky parts back, and waiting a bit longer before
deliberately breaking backwards compatibility and updating requirements.

When doing it the old way, latest setuptools will emit warnings and give a
deadline at 2026-Feb-18. It might thus be a good compromise to redo the change
for 7.2 .

--HG--
branch : stable
extra : rebase_source : bc08b797f80fb949260fb56e767f69ab74d5ae01
extra : topic : mads-pyproject-license