rust-dependencies: upgrade `indicatif`
This is a new cycle, so we should upgrade our dependencies.

This is done in a separate change because it's technically incompatible
(in terms of semantic versionning).
In practice it doesn't change anything for us.

--HG--
extra : rebase_source : d7c741b6f6264d9d3d84b90e5c159a337fd65609
extra : topic : rust-early-cycle