rust-dependencies: upgrade `toml`
This is a new cycle, so we should upgrade our dependencies.

This is done in a separate change because it's technically incompatible
(in terms of semantic versionning). This was a full rewrite of the crate
to make it more performant, more expressive in terms of errors, etc.

In practice it doesn't change anything for us.

--HG--
extra : rebase_source : 8bb2f98bb5c5bde6f5720e5570ec55cf437c2f73
extra : topic : rust-early-cycle