hgweb: validate plain filesystem paths exist in `--web-conf` file
Since walking the directory tree to find repositories in `scmutil.walkrepos()`
will implicitly validate that paths exist when it stats the directory path, we
should make sure these plain paths exist too.  Otherwise accessing them result
in a cryptic 500 on the client, and no other info unless they have access to the
server error log.

The error message will be different there vs here (there will raise a native
`FileNotFoundException`, which contains the Win32 error code, for example), but
this is a little nicer because we can give a little more context about what
wasn't found.

An existing test that goes back to 8dfdf0219494 in 2010 seems to indicate that
the intention was to silently ignore missing repos.  But I can't tell if the
test was broken, because the headers are fetched for the root path instead of
the missing repo path.  So that test is adjusted to clear the bogus paths ASAP,
and not have a cascading failure.  While technically this may be a behavior
change, I think it's more reasonable that if the permissive search for repos
with globs case strictly checks that a path exists, the explicit path case will
also check.  I don't think anyone will complain that configuration issues are
easier to diagnose.

--HG--
extra : __touch-noise__ : 4125387285
extra : rebase_source : 5b731661b96d8cf899ca3f6af6a9ed3f71630072
extra : topic : webconf-path-validation