test-filelog: only open writable filelog in transaction
This changes test-filelog.py to only create a filelog as writable while it holds
the transaction. This is necessary for an upcoming change with the file index
which will assert that there is a transaction when using writable=True.

--HG--
extra : rebase_source : 83d34bf555cde103a07d8fecb25433f6f1473ce4
extra : topic : file-index.ready