revlogutils: fix writable argument in repair_issue6528
The main function `repair_issue6528` was not taking `to_report` into account,
and the helper function `_from_report` was not taking `dry_run` into account,
when setting the `writable` argument. This makes no difference now, but it will
when we start asserting that a transaction exists whenever accessing a filelog
with writable=True.

--HG--
extra : topic : file-index.ready