typing: add some typehints to `mercurial.ui.httppasswordmgrdbproxy`
This is the easy stuff, so we can get a partial protocol class.  The hints are
courtesy of PyCharm with Python 3.11, and minor additions to also allow bytes,
and disallow None for the `realm` (because of the typing on the underlying
conversion function).

--HG--
extra : rebase_source : d8cbe1a8567c251637415e95ddfe12f4d4564528
extra : topic : typing-ui-stub