typing: add some typehints to `mercurial.ui.ui`
There's plenty more to do (including making a protocol class for the
progressbar), but that can be done later.  This is low hanging fruit before some
functions are copied off to a protocol class, in order to prevent regressions
around things like `ui.status()`, now that various things formerly typed with
this class are now typed to `UiT` (which is aliased to `Any`).

The cast of the `_unset` sentinel value assigned to the `default` arg of
`configbool` is a bit hokey, but it's specifically compared by its identity, and
the cast prevents pytype from thinking that this function returns an `object`,
even though that case is specifically handled.  I suppose this arg could also
accept `None` since the code explicitly detects and returns that, but let's see
how this fairs.  As it is, the return type isn't specified because it can return
`None`, and it's a pain to cast away that value everywhere this method is
called.  Right now, pytype types the return as `Any`.

--HG--
extra : rebase_source : 34cec6e7c3a41d27392e6bc59edc6f549bcb1c59
extra : topic : typing-ui-stub