osutil: avoid Py_GetArgcArgv in setprocname
NOTE: This change was taken from Meta's Sapling repository on GitHub[^1].

Original description:

  > Avoid using `Py_GetArgcArgv` to get argv. The Python interpreter may not even
  > have the real argv.
  >
  > Read backwards from `environ` on Linux. Use `_NSGetArgv()` on macOS.

---

The `osutil` module provides a `setprocname` function, which is currently used
by chg to set the process name of worker processes, which helps with debugging.

Support for `setprocname` on Linux and MacOS was dropped[^1] from upstream
Mercurial because it relies on CPython's `Py_GetArcArgv` function, which was
changed in Python 3.x to return a pointer to a copy of the argv, thereby
preventing rewriting the process's command line. This limitation was noted
at the time `setprocname` was introduced[^2].

In Sapling (Meta's fork of Mercurial), rather than removing this code, it was
instead updated to work correctly with Python 3. This revision copies over the
initial version from the Sapling repo. Later updates to this code in the Sapling
repo will be copied over in later commits in this stack.

[^1]: https://github.com/facebook/sapling/commit/9221225890b55e20593f6a83bcdd43cbfb952ebe

[^2]: See revision 3aa1b7ded52cb27e0e28c8e198bfd516332e887d.
      ("cext: remove inline rewriting of argv")

[^3]: See revision 0852161588c6e7af9c62a5582ebe95f7d23c2938.
      ("osutil: implement setprocname to set process title for some platforms")

--HG--
extra : rebase_source : 3c84afc3e9c4e0908ea210f83283da81631fa54a
extra : topic : restore-setprocname-useargvwrite