ci: add more sink for the linux test
Adding sharding "increased" the size of the job that other sink rely on. So we
need more sink. We also fix the sinks to actually wait on all tests and not just
the c one. Even if the wheel only contains the c for now, failure elsewhere
might reveal a wider issue. In addition we might start to use a different
variant than the c in the wheel at some point.

--HG--
extra : rebase_source : 18bc4b653379540403255a77f8a38c4bf5ec6ca4
extra : topic : more-sink