debugshell: allow the script to be multiline
This changes the behavior of `debugshell` to support multiline scripts.

The unintentional side-effect is that the final value produced
by the script is not printed anymore.

None of the  `debugshell` snippets that I've seen relied on the
implicit printing, but it's something to watch out for.

Pierre-Yves argues small breakage is fine for a command like this.

--HG--
extra : rebase_source : 9234d521b32591a2e8632ce0eadeb2c8e129b7cd
extra : topic : debugshell-multiline