lfs: avoid wrapping `makefilestorage`
The work done here is very small and could be done in core directly. So we do it
in core directly to remove on user of the weird `makefilestorage` function.

We could also setup another way to wrap this feature injection, but it does not
seems important enough to do so.

--HG--
extra : rebase_source : 0b0f49a5290b1451b23dcae3343ffe6b9c2ee08d
extra : topic : cleanup.repo-init