changegroup: type the `_stream` attribute the best we could
This "stream" attribute is tricky because it is very often a very basic
"stream" with only a `read` argument. However in some case there are full file
that we can seek and need to close (for example when using bundle repo).

The current architecture doesn't easily allow to distinct the two cases and
improving typing annotation will have `pytype` chocking on this inconsistency.
So for now, we put our head in the sand and pretend all stream are full
featured from a typing point of view.

This will need revisiting later.

--HG--
extra : rebase_source : 80d25467ffcbeff7a553445b92e323699ad0c37e
extra : topic : typing.compression