delta-search: introduce a revlog.reuse-external-delta-compression config
This config allow to directly reuse the incoming compressed delta when
applicable.

Support for actually exchanging such compressed delta over the wire still needs
to be implemented in the changegroup-v4 format.

The shape of the feature and associated config might change before the release,
but this the core of it is now implemented.

--HG--
extra : rebase_source : 5e899f68dbe121d1e33a54d3cc15c0877f527134
extra : topic : compressed-delta-exchange