changegroup: move `emitrevisions` back on the changelog class
This logic was split out in `storageutils` a while back with the project of
making more abstract storage when revlog for it. However no such things
actually happened and the function became even more symbiotic with the revlog
class making the whole interface weird and complex to read. Trying to add
typing information in that state would be be midly insane.

So I am moving back this method to the revlog itself. I am not super fan of
adding more code to this already gigantic class and module. However that's a
lesser concerns that the current complexity we get from having it separated.
And we can deal with that problem later.

We will spend the next few changesets to clean up the API so that we can
actually get a benefit from this move.

--HG--
extra : rebase_source : f6fa640a757a749f2a445fe781c2e15711f39718
extra : topic : inline-emit-revision