ci: update comment about what we test on macOS
This isn't great, because it's actually controlled by a symlink named `python`
on the system, to point to the `python3.Y` that should actually be used.  But,
we might as well be accurate, and it's valid reasoning to build all of the
wheels.

--HG--
branch : stable
extra : rebase_source : 1f3e9fac97b87e48529da13e3eedd47f053bf99b
extra : topic : ci-macos-always-on