verify: warnorphanstorefiles only for fncache
This makes the hg verify warnorphanstorefiles logic specific to the fncache.
Previously it tried to be agnostic of the store, and just check if there were
any discrepancy between the store's files and filelogs found in manifests.
However, the file index tracks filelogs at higher level than the fncache, so it
will be better to implement verification logic for it separately. This changeset
is in preparation for that.

--HG--
extra : rebase_source : fdd97ea5133184d61b886e4b2aa7307f36d2c193
extra : topic : file-index