lfs: add a debug command to display information about the local blob cache
Someday we can add more fancy stuff like file count, sizes, etc.  For now, I'm
sick of having to look in the code and trapse through all of the conditionals
and default values and overriding config options whenever I need to know the
location for a platform.

Some interesting stuff I learned when testing:

  - The vfs object doesn't expand or realpath() a relative directory from the
    `lfs.usercache` config option (we can do this without breaking largefiles)

  - The platform default path isn't inside a 'Mercurial' directory, rather it's
    something like `C:\Users\Matt\AppData\Local\lfs` (we can't fix this without
    either being inconsistent with largefiles, or breaking largefiles)

I'm not sure if there's prior art for a path that may or may not be present.  If
templated, it makes more sense to have an empty path when the path is missing,
to make it easier to handle programmatically.  But if not templated, "path: "
just looks like the info is missing.  So we split the difference, and can fix it
later if needed, because it's a debug command.

--HG--
extra : rebase_source : 9d188081d0feaea57ab66038852d0374c62e26a1
extra : topic : lfs-convert