fncache: explicitly register relevant files in a few spots
There are only a few file that needs to be added to the fncache: the `.i` and
`.d` of filelog and the `.i` and `.d` of non-root manifest, and we can precisely
detect when they are created from within the revlog (and upgrade) code itself.

So instead of having code within the vfs doing on-the-fly detection of the
creation of such file for every file operations… We could just explicitly
register them a these few key moment.

To some extend, this "obvious" simplification have been in front of us since
forever… However it is much easier to see it was obvious once that option was
noticed.

This open the way to moving the encoding of path at the store level (when we
compute the revlog radix) in later changesets. Another big simplification.

--HG--
extra : rebase_source : 222096ebed9985881a07887826f8a18bc162cb6e
extra : topic : explicit-fncache