rust-manifest: quickly compare identical sections with SMID
When many line are the same, we can quickly walk over them by direct bytes
comparison. We nudge the compiler into using vectorized instruction for that.

When a difference is detected, we fallback to comparing line themselves, until
we start noticing that all the content seems the same again.

This greatly speeds the manifest diffing into pretty much nothing.

For example computing delta's between 1000 manifest on mozilla central is now
twice as fast, and now 80% of the time is spent restoring the manifest and as
little as about ~15% spent in the diffing itself. It was 35% for building the
full text and 55% for diffing before we implemented the diffing in Rust.

Here is a sample benchmark for a real life usecase:

### data-env-vars.name                                        = netbeans-2018-08-01-zstd-sparse-revlog
  # benchmark.name                                            = hg.command.unbundle
  # bin-env-vars.hg.flavor                                    = rust
  # bin-env-vars.hg.py-re2-module                             = default
  # benchmark.variants.issue6528                              = disabled
  # benchmark.variants.rebundle                               = no
  # benchmark.variants.resource-usage                         = high
  # benchmark.variants.reuse-external-delta                   = yes
  # benchmark.variants.reuse-external-delta-parent            = no
  # benchmark.variants.reuse-external-suspicious-delta-parent = default
  # benchmark.variants.revs                                   = any-1000-extra-rev
  # benchmark.variants.source                                 = unbundle
  # benchmark.variants.validate                               = no
  # benchmark.variants.verbosity                              = quiet
mpatch.diff:        8.460202
manifest-diff:      6.337972  (-25.08%, -2.12)
manifest-diff+simd: 5.466991  (-35.38%, -2.99)

--HG--
extra : rebase_source : f9955f19b4202cff12f59d24c165aa087677e4ec
extra : topic : manifest-diff