streamclone: refactor _copy_files
This refactors the _copy_files function to take the list of entries directly,
rather than taking a flattened list of files. The motivation for this is to make
it work with the file index, which I will do in a follow-up change.

--HG--
extra : rebase_source : 1a048d58b1c44d49b5e6858180448eaca843b850
extra : topic : file-index.clone