exewrapper: simplify the script search on Windows
Before, the logic was roughly:

    - Take full executable path, chop off '.exe', and see if that python script
      existed.  If so, that's the script to run.
    - If it didn't, append "exe.py" to the truncated path (this likely intended
      a dot before the "exe", but who knows), and then use that as the script.
    - In all cases, chop the last path component off, and use that as PYTHONHOME
      (but only with HackableMercurial)

That doesn't work for the case of `make local` creating `hg-local.exe`, because
there's neither `hg-local` nor `hg-localexe.py`.  So now we just assume the
script to run is `hg`, next to `hg.exe`.

Note that when run from either the root of the repo, or in the private venv,
`hg debuginstall` both point to the system python for "Python lib", and the
repository for "installed modules" (instead of the venv).  But that's how it
was before too.

--HG--
extra : absorb_source : 4d4dfee627304861d05b61389e4b8678078758f4
extra : rebase_source : d531a96267446554e55027f5978d75e65e5242d5
extra : topic : windows-build-fixes