make-local: adjust to work on Windows again
The first problem was the `-p` to `uv` for the python executable requires the
'.exe' suffix.

The second problem was the hardcoded "bin" reference- it's "Scripts" in the venv
on Windows.

The third problem was if the symlink points to `hg` instead of `hg.exe` in the
venv, then the python script is linked in the root of the repo, and MSYS "runs"
that by looking at the shebang line (which will point to the venv python), and
invoke it directly.  That doesn't set the legacy stdio option, so any command
that involves the pager will have its output lost.

The symlink is dubious here (`ls -l` doesn't show it as a symlink, and the size
in the src and dest match, so it may have just done a hardlink), but the net
result is that it appears to work.

--HG--
extra : absorb_source : c76136376f42f83858338b0ece6c1c76d15b2147
extra : amend_source : e41a0745df776e6e338960ccf473e1ff1ed31511
extra : rebase_source : 804fb38c8672a9e10449f2d76d8ead3f7211a382
extra : topic : windows-build-fixes