changegroup: turn the "01" packer into a `cgpacker` subclass
For now, the sub-classing does do much, but we will be able to clean more things
when all of them are converted.

While the previous setup based on a small call back function kind of work, it
is becoming more and more impractical in terms of extensibility and typing.

--HG--
extra : rebase_source : b68b2d3112bde9273912616192489fc5515cf2bc
extra : topic : changegroup-typing