docket: reserve all 0s as UNSET_UID
This adds UNSET_UID = b"00000000" and ensures that make_uid never returns it.
The motivation for this is it will be useful in the file index to use the
UNSET_ID as an indicator that we haven't yet written a file.

--HG--
extra : rebase_source : e62dbf4d0f15e8c16a6d61e9e978aa6e7ce93450
extra : topic : file-index.remove